Moved GeographicLib classes and package
