+++ sdk/sdk-api/src/main/java/mil/emp3/api/utils/EllipticFunction.java
@@ -1,4 +1,8 @@
-package net.sf.geographiclib;
+package mil.emp3.api.utils;
+
+import net.sf.geographiclib.GeoMath;
+import net.sf.geographiclib.GeographicErr;
+import net.sf.geographiclib.Pair;
 
 import static java.lang.Math.*;
